[ACS-5343] refactor Playwright framework for ACA
