+++ .github/workflows/pull-request.yml
@@ -180,6 +180,14 @@ jobs:
     needs: [lint, build, unit-tests]
     name: 'E2e test suites: Folder Rules - Playwright'
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        e2e-suites:
+          - name: "actions"
+            id: 1
+          - name: "folder-rules"
+            id: 2
     steps:
       - name: Checkout
         uses: actions/checkout@v3
@@ -206,7 +214,7 @@ jobs:
         run: npx playwright install chromium
       - uses: ./.github/actions/run-e2e
         with:
-          options: "e2e/playwright/tests/folder-rules/playwright.config.ts"
+          options: "playwright.config.ts tests/${{ matrix.e2e-suites.name }}"
           artifact-name: folder-rules
           test-runner: playwright
       - uses: ./.github/actions/after-e2e
@@ -241,5 +249,3 @@ jobs:
             echo -e "\e[31mPRs are not mergeable with conditional build. This build was run with custom ADF branch: $BRANCH \e[0m"
             exit 1
           fi;
-
-
