+++ e2e/playwright/src/page-objects/components/conditions.component.ts
