+++ e2e/playwright/src/page-objects/components/data-table/index.ts
@@ -22,23 +22,4 @@
  * from Hyland Software. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { BasePage, ToolbarComponent } from '@alfresco/playwright-shared';
-import { Page } from '@playwright/test';
-import { ManageRulesDialogComponent } from '../components/manage-rules-dialog.component';
-import { ActionsDropdownComponent } from '../components/actions-dropdown.component';
-import { ConditionComponent } from '../components/conditions.component';
-import { ManageRules } from '../components/manage-rules.component';
-
-export class NodesPage extends BasePage {
-  private static pageUrl = 'nodes';
-
-  constructor(page: Page) {
-    super(page, NodesPage.pageUrl);
-  }
-
-  public toolbar = new ToolbarComponent(this.page);
-  public manageRulesDialog = new ManageRulesDialogComponent(this.page);
-  public actionsDropdown = new ActionsDropdownComponent(this.page);
-  public conditionsDropdown = new ConditionComponent(this.page);
-  public manageRules = new ManageRules(this.page);
-}
+export * from './data-table.component';
