+++ e2e/playwright/src/page-objects/components/manage-rules.component.ts
@@ -22,29 +22,16 @@
  * from Hyland Software. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { BaseComponent } from '@alfresco/playwright-shared';
-import { Page } from '@playwright/test';
+import { Locator, Page } from '@playwright/test';
+import { BaseComponent } from 'e2e/playwright/src';
 
-export class ManageRulesDialogComponent extends BaseComponent {
-  private static rootElement = 'aca-edit-rule-dialog';
+export class ManageRules extends BaseComponent {
+  private static rootElement = '.aca-manage-rules';
 
-  public createRuleButton = this.getChild('[data-automation-id="edit-rule-dialog-submit"]');
-  private ruleNameInputLocator = this.getChild('[id="rule-details-name-input"]');
-  public addConditionButton = this.getChild('[data-automation-id="add-condition-button"]');
-  public addConditionGroupButton = this.getChild('[data-automation-id="add-group-button"]');
-  public fieldDropDown = this.getChild('[data-automation-id="field-select"]');
-  public comparatorDropDown = this.getChild('[data-automation-id="comparator-select"]');
-  private valueField = this.getChild('[data-automation-id="value-input"]');
+  public getGroupsList = (optionName: string): Locator => this.getChild('.aca-rule-list-item__header', { hasText: optionName });
+  public disableRuleToggle = this.getChild('.aca-manage-rules__container .mat-slide-toggle-bar').first();
 
   constructor(page: Page) {
-    super(page, ManageRulesDialogComponent.rootElement);
-  }
-
-  async typeRuleName(ruleName: string): Promise<void> {
-    await this.ruleNameInputLocator.type(ruleName, { delay: 50 });
-  }
-
-  async typeConditionValue(ruleName: string, index: number): Promise<void> {
-    await this.valueField.nth(index).type(ruleName, { delay: 50 });
+    super(page, ManageRules.rootElement);
   }
 }
