+++ e2e/playwright/src/page-objects/components/spinner.component.ts
