+++ e2e/playwright/src/page-objects/global-variables.ts
