+++ e2e/playwright/src/page-objects/pages/personal-files.page.ts
@@ -22,6 +22,24 @@
  * from Hyland Software. If not, see <http://www.gnu.org/licenses/>.
  */
 
-export * from './data-table.component';
-export * from './mat-menu.component';
-export * from './toolbar.component';
+import { BasePage, DataTableComponent } from 'e2e/playwright/src';
+import { Page } from '@playwright/test';
+import { AcaPageLayoutHeader } from '../components/aca-page-layout-header.component';
+import { MatMenuComponent } from '../components/mat-menu.component';
+import { AdfFolderDialogComponent } from '../components/adf-folder-dialog.component';
+
+export class PersonalFilesPage extends BasePage {
+  private static pageUrl = 'personal-files';
+
+  constructor(page: Page) {
+    super(page, PersonalFilesPage.pageUrl);
+  }
+
+  public acaHeader = new AcaPageLayoutHeader(this.page);
+
+  public matMenu = new MatMenuComponent(this.page);
+
+  public folderDialog = new AdfFolderDialogComponent(this.page);
+
+  public dataTable = new DataTableComponent(this.page);
+}
